add write_to_file HTK format for FeatureVector class

+++ action_recognition/include/action_recognition/Vector3D.hpp
@@ -8,24 +8,24 @@
 class Vector3D{
 
 private:
-  std::vector<double> vector3D_;
+  std::vector<float> vector3D_;
 
 public:
   /* Constructor for an emtpy 3D vector*/
   Vector3D();
   /*Constructor for a 3D vector*/
-  Vector3D(double x, double y, double z);
-  Vector3D(std::vector<double> vector3D);
+  Vector3D(float x, float y, float z);
+  Vector3D(std::vector<float> vector3D);
 
-  void set_new_values(double x, double y, double z);
+  void set_new_values(float x, float y, float z);
 
-  double get_x(void);
-  double get_y(void);
-  double get_z(void);
+  float get_x(void);
+  float get_y(void);
+  float get_z(void);
 
-  void set_x(double x);
-  void set_y(double y);
-  void set_z(double z);
+  void set_x(float x);
+  void set_y(float y);
+  void set_z(float z);
 
   Vector3D normalize(void);
 };
